src/robotic_arm_control/requirements.txt
# 基础依赖（保留你指定的版本）
numpy>=1.24.0
pyyaml>=6.0
pyqt5>=5.15.9
pyopengl>=3.1.6
matplotlib>=3.7.1

# 数值优化依赖
numba>=0.57.0  # JIT编译加速

# MuJoCo 相关依赖
mujoco>=2.3.7
glfw>=2.5.5  # MuJoCo 渲染依赖